# api-gateway/openapi.yaml
openapi: 3.1.0
info:
  title: EcoSignal API Gateway
  description: >
    Public API for EcoSignal — connects Italian citizens' daily habits
    to real local environmental data, generating personalized climate
    narratives and actionable CO₂-reduction suggestions.
  version: 1.0.0
  contact:
    name: EcoSignal Team
servers:
  - url: http://localhost:8000
    description: Local development
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    TransportMode:
      type: string
      enum: [car, transit, bike, walk, mixed]
    DietType:
      type: string
      enum: [meat_daily, meat_weekly, vegetarian, vegan]
    HomeType:
      type: string
      enum: [apartment, house]
    AqiLabel:
      type: string
      enum: [good, moderate, unhealthy, hazardous]
    ActionCategory:
      type: string
      enum: [transport, food, home, community]
    Effort:
      type: string
      enum: [easy, medium, hard]

    # ── Auth ───────────────────────────────────────────────
    RegisterRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          maxLength: 254
        password:
          type: string
          format: password
          minLength: 8
          maxLength: 128
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
    RefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
    TokenResponse:
      type: object
      required: [access_token, refresh_token, token_type, expires_in]
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
          example: bearer
        expires_in:
          type: integer
          description: Access token lifetime in seconds
          example: 3600

    # ── User Profile / Onboarding ─────────────────────────
    OnboardingRequest:
      type: object
      required: [zip_code, transport_mode, diet_type, home_type, home_size_sqm]
      properties:
        zip_code:
          type: string
          pattern: '^\d{5}$'
          description: Italian CAP (5 digits)
          example: "20121"
        transport_mode:
          $ref: '#/components/schemas/TransportMode'
        diet_type:
          $ref: '#/components/schemas/DietType'
        home_type:
          $ref: '#/components/schemas/HomeType'
        home_size_sqm:
          type: integer
          minimum: 10
          maximum: 500
          example: 75
    ProfileUpdateRequest:
      type: object
      minProperties: 1
      properties:
        zip_code:
          type: string
          pattern: '^\d{5}$'
        transport_mode:
          $ref: '#/components/schemas/TransportMode'
        diet_type:
          $ref: '#/components/schemas/DietType'
        home_type:
          $ref: '#/components/schemas/HomeType'
        home_size_sqm:
          type: integer
          minimum: 10
          maximum: 500
    HabitProfile:
      type: object
      required:
        - user_id
        - zip_code
        - transport_mode
        - diet_type
        - home_type
        - home_size_sqm
        - created_at
        - updated_at
      properties:
        user_id:
          type: string
          format: uuid
        zip_code:
          type: string
          pattern: '^\d{5}$'
        transport_mode:
          $ref: '#/components/schemas/TransportMode'
        diet_type:
          $ref: '#/components/schemas/DietType'
        home_type:
          $ref: '#/components/schemas/HomeType'
        home_size_sqm:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    # ── Footprint ─────────────────────────────────────────
    FootprintResult:
      type: object
      required:
        - user_id
        - week_start
        - total_kg_co2
        - breakdown
        - vs_national_avg_pct
      properties:
        user_id:
          type: string
          format: uuid
        week_start:
          type: string
          format: date
        total_kg_co2:
          type: number
          format: float
          example: 26.7
        breakdown:
          type: object
          required: [transport, food, home]
          properties:
            transport:
              type: number
              format: float
            food:
              type: number
              format: float
            home:
              type: number
              format: float
          additionalProperties: false
        vs_national_avg_pct:
          type: number
          format: float
          description: >
            Percentage difference vs national average.
            Negative means below average (better).
          example: -0.12
    FootprintHistory:
      type: object
      required: [user_id, weeks]
      properties:
        user_id:
          type: string
          format: uuid
        weeks:
          type: array
          items:
            $ref: '#/components/schemas/FootprintResult'
          maxItems: 8

    # ── Environmental Data ────────────────────────────────
    AirQualityReading:
      type: object
      required:
        - zip_code
        - station_id
        - aqi_label
        - fetched_at
        - source
      properties:
        zip_code:
          type: string
        station_id:
          type: string
        pm25:
          type: ['number', 'null']
          format: float
        pm10:
          type: ['number', 'null']
          format: float
        o3:
          type: ['number', 'null']
          format: float
        aqi_label:
          $ref: '#/components/schemas/AqiLabel'
        fetched_at:
          type: string
          format: date-time
        source:
          type: string
          enum: [openaq, arpa, fallback]
    ClimateContext:
      type: object
      required: [zip_code, summary, fetched_at]
      properties:
        zip_code:
          type: string
        summary:
          type: string
          description: >
            Human-readable sentence describing today's local climate
            anomaly relative to historical norms.
          example: >
            Oggi a Milano la temperatura è 3°C sopra la media stagionale,
            con livelli di umidità nella norma.
        fetched_at:
          type: string
          format: date-time

    # ── Narrative ─────────────────────────────────────────
    NarrativeResponse:
      type: object
      required:
        - user_id
        - headline
        - body
        - local_context
        - source_note
        - generated_at
      properties:
        user_id:
          type: string
          format: uuid
        headline:
          type: string
          maxLength: 120
        body:
          type: string
        local_context:
          type: string
        source_note:
          type: string
          example: "Based on ISPRA 2024 · OpenAQ today"
        generated_at:
          type: string
          format: date-time

    # ── Actions ───────────────────────────────────────────
    ActionSuggestion:
      type: object
      required:
        - action_id
        - title
        - description
        - category
        - co2_saving_kg_week
        - effort
        - completed
      properties:
        action_id:
          type: string
        title:
          type: string
        description:
          type: string
        category:
          $ref: '#/components/schemas/ActionCategory'
        co2_saving_kg_week:
          type: number
          format: float
        effort:
          $ref: '#/components/schemas/Effort'
        completed:
          type: boolean
    ActionCompleteResponse:
      type: object
      required: [action_id, completed, completed_at]
      properties:
        action_id:
          type: string
        completed:
          type: boolean
          example: true
        completed_at:
          type: string
          format: date-time

    # ── Community ─────────────────────────────────────────
    NeighborhoodStats:
      type: object
      required:
        - zip_code
        - active_users
        - avg_kg_co2_week
        - top_action
        - current_challenge
      properties:
        zip_code:
          type: string
        active_users:
          type: integer
          description: Anonymized count of active users in this CAP
        avg_kg_co2_week:
          type: number
          format: float
          description: Average weekly footprint for this CAP
        top_action:
          type: string
          description: Most completed action title in this CAP
        current_challenge:
          $ref: '#/components/schemas/WeeklyChallenge'
    WeeklyChallenge:
      type: object
      required:
        - challenge_id
        - title
        - description
        - week_start
        - week_end
        - participants
      properties:
        challenge_id:
          type: string
        title:
          type: string
        description:
          type: string
        week_start:
          type: string
          format: date
        week_end:
          type: string
          format: date
        participants:
          type: integer
    ChallengeJoinResponse:
      type: object
      required: [challenge_id, joined, joined_at]
      properties:
        challenge_id:
          type: string
        joined:
          type: boolean
          example: true
        joined_at:
          type: string
          format: date-time

    # ── Health ────────────────────────────────────────────
...
    ErrorResponse:
      type: object
      required: [detail]
      properties:
        detail:
          type: string
          description: Human-readable error message
          example: Invalid or expired token
paths:
...
  /api/v1/auth/register:
    post:
      tags: [auth]
      operationId: register
      summary: Create a new user account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User created, tokens returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          description: Validation error (e.g. weak password)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/login:
    post:
      tags: [auth]
      operationId: login
      summary: Authenticate and obtain tokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/refresh:
    post:
      tags: [auth]
      operationId: refreshToken
      summary: Refresh an expired access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: New token pair returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ── Profile ───────────────────────────────────────────
  /api/v1/onboarding:
    post:
      tags: [profile]
      operationId: submitOnboarding
      summary: Submit initial habit profile during onboarding
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OnboardingRequest'
      responses:
        '201':
          description: Profile created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HabitProfile'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Profile already exists for this user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/profile:
    get:
      tags: [profile]
      operationId: getProfile
      summary: Get current user's habit profile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Profile returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HabitProfile'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Profile not found — onboarding not completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    patch:
      tags: [profile]
      operationId: updateProfile
      summary: Update one or more habit fields
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfileUpdateRequest'
      responses:
        '200':
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HabitProfile'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ── Footprint ─────────────────────────────────────────
  /api/v1/footprint/calculate:
    post:
      tags: [footprint]
      operationId: calculateFootprint
      summary: Trigger CO₂ footprint calculation for the current week
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Calculation result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FootprintResult'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Profile incomplete — complete onboarding first
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal calculation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/footprint/history:
    get:
      tags: [footprint]
      operationId: getFootprintHistory
      summary: Get weekly footprint trend (last 8 weeks)
      security:
        - bearerAuth: []
      parameters:
        - name: weeks
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 8
            default: 8
          description: Number of past weeks to return (max 8)
      responses:
        '200':
          description: Weekly history returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FootprintHistory'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ── Environmental Data ────────────────────────────────
  /api/v1/env/air-quality:
    get:
      tags: [env]
      operationId: getAirQuality
      summary: Get current air quality reading for the user's ZIP code
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Air quality data returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AirQualityReading'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: All external data sources unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/env/climate-context:
    get:
      tags: [env]
      operationId: getClimateContext
      summary: Get local climate anomaly summary for user's ZIP code
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Climate context returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClimateContext'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Climate data unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ── Narrative ─────────────────────────────────────────
  /api/v1/narrative:
    get:
      tags: [narrative]
      operationId: getNarrative
      summary: >
        Generate or return a cached personalized climate narrative
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Narrative returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NarrativeResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: >
            Cannot generate narrative — profile or footprint data missing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Narrative generation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ── Actions ───────────────────────────────────────────
  /api/v1/actions:
    get:
      tags: [actions]
      operationId: getActions
      summary: Get top 3 ranked action suggestions for the user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Ranked action list returned
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ActionSuggestion'
                maxItems: 3
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/actions/{action_id}/complete:
    post:
      tags: [actions]
      operationId: completeAction
      summary: Mark an action as completed
      security:
        - bearerAuth: []
      parameters:
        - name: action_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Action marked as completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionCompleteResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Action not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ── Community ─────────────────────────────────────────
  /api/v1/community/stats:
    get:
      tags: [community]
      operationId: getCommunityStats
      summary: Get anonymized neighborhood stats for user's ZIP
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Neighborhood stats returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NeighborhoodStats'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/community/challenge/join:
    post:
      tags: [community]
      operationId: joinChallenge
      summary: Join the current weekly challenge
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Successfully joined
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChallengeJoinResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Already joined or no active challenge
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ── Health ────────────────────────────────────────────
...